Make the following changes:
,

Find this element:

  <Alert variant="destructive" className="my-8 shadow-md">

with the text content:

    ErrorGeolocation error: User denied Geolocation

Relevant files:

    - https://6000-firebase-studio-1746577755238.cluster-c23mj7ubf5fxwq6nrbev4ugaxa.cloudworkstations.dev/_next/static/chunks/src_83eaa1ae._.js
    - https://6000-firebase-studio-1746577755238.cluster-c23mj7ubf5fxwq6nrbev4ugaxa.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1746577755238.cluster-c23mj7ubf5fxwq6nrbev4ugaxa.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

rectify this error. If permission is required, let the app ask for the required permission from the user. If the permission is denied, then the display the details from the Location asked in the "Enter City Name" window.